Update lot of patterns code

+++ Flyweight/FactoryModel.cpp
@@ -1,6 +1,6 @@
 ﻿#include "FactoryModel.h"
 
-size_t FactoryModel::GetSize() const
+uint32_t FactoryModel::GetSize() const
 {
     return m_models.size();
 }
